sql: allow ON CONFLICT REPLACE only for PK index
The problem is that constraints which require VDBE bytecode
will be executed before making insertion into Tarantool. For
three major constraints - UNIQUE, FOREIGN-KEY, CHECK the
execution order doesn't matter if error action is not REPLACE,
because one constraint violation doesn't affect others.

But violation of constraint with ON CONFLICT REPLACE leads to a
removal of whole tuple. When only PRIMARY KEY has REPLACE as
error action, it is not a problem and such behavior is absolutely
similar to Tarantool. But when we are dealing with secondary index
/NOT NULL constraint with ON CONFLICT REPLACE, conflict in
secondary index leads to a whole tuple removal, and as a result
when there is also a conflict in primary key index, the insertion
will lead to success instead of raising error.

In this patch possibility to create constraint non PRIMARY KEY
constraint with ON CONFLICT REPLACE was banned, that kind of
behavior will lead to a proper error message.

Closes #2963